add SimCoordinator module; move postprocess to Plotting
