+++ docs/make.jl
@@ -1,17 +1,6 @@
 using ClimaCoupler
 using Documenter, Literate
 using CairoMakie, ClimaCoreMakie, GeoMakie, Makie, Poppler_jll, Printf, Oceananigans
-using ClimaCoupler:
-    Checkpointer,
-    ConservationChecker,
-    FieldExchanger,
-    FluxCalculator,
-    Input,
-    Interfacer,
-    SimOutput,
-    TimeManager,
-    Utilities,
-    Plotting
 
 const COUPLER_DIR = joinpath(@__DIR__, "..")
 const EXPERIMENTS_DIR = joinpath(@__DIR__, "..", "experiments")
@@ -81,6 +70,7 @@ interface_pages = [
     "fieldexchanger.md",
     "fluxcalculator.md",
     "interfacer.md",
+    "simcoordinator.md",
     "timemanager.md",
     "utilities.md",
     "simoutput.md",
